app/higuchi_pr/requirements.txt
django==5.2.1
gunicorn==23.0.0
djangorestframework==3.16.0
requests==2.25.1